requirements.txt
APScheduler==3.11.0
Flask==2.0.3
Flask-APScheduler==1.12.3
requests==2.28.1
psycopg2-binary==2.9.5
elasticsearch==8.10.1
Werkzeug==2.0.3
flask_cors